requirements.txt
phidata x
python-dotenv
yfinance 
duckduckgo-search
fastapi 
uvicorn 
groq
packaging 
requests